+++ include/protowire/pxf.h
@@ -26,6 +26,7 @@
 #include "protowire/pxf/options.h"
 #include "protowire/pxf/parser.h"  // Document, Parse
 #include "protowire/pxf/result.h"
+#include "protowire/pxf/schema.h"  // ValidateDescriptor, Violation
 
 namespace protowire::pxf {